requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pypdf2==3.0.1
langchain==0.2.5
langchain-google-genai==1.0.5
python-dotenv==1.0.1
streamlit==1.36.0
faiss-cpu==1.8.0
sentence-transformers==3.0.1